chore: separate http handlers in smaller files
