+++ internal/api/handlers_certificate_requests_test.go
@@ -6,13 +6,11 @@ import (
 	"log"
 	"net/http"
 	"net/http/httptest"
-	"regexp"
 	"strings"
 	"testing"
 
 	server "github.com/canonical/notary/internal/api"
 	"github.com/canonical/notary/internal/db"
-	"github.com/golang-jwt/jwt"
 )
 
 const (
@@ -138,18 +136,6 @@ var (
 	expectedGetCertReqResponseBody2  = fmt.Sprintf("{\"id\":4,\"csr\":\"%s\",\"certificate\":\"\"}", trimmed(AppleCSR))
 	expectedGetCertReqResponseBody3  = fmt.Sprintf("{\"id\":2,\"csr\":\"%s\",\"certificate\":\"%s\\n%s\"}", trimmed(BananaCSR), trimmed(BananaCert), trimmed(IssuerCert))
 	expectedGetCertReqResponseBody4  = fmt.Sprintf("{\"id\":2,\"csr\":\"%s\",\"certificate\":\"\"}", trimmed(BananaCSR))
-)
-
-const (
-	adminUser              = `{"username": "testadmin", "password": "Admin123"}`
-	validUser              = `{"username": "testuser", "password": "userPass!"}`
-	invalidUser            = `{"username": "", "password": ""}`
-	noPasswordUser         = `{"username": "nopass"}`
-	adminUserNewPassword   = `{"id": 1, "password": "newPassword1"}`
-	userNewInvalidPassword = `{"id": 1, "password": "password"}`
-	userMissingPassword    = `{"id": 1}`
-	adminUserWrongPass     = `{"username": "testadmin", "password": "wrongpass"}`
-	notExistingUser        = `{"username": "not_existing", "password": "user"}`
 )
 
 func TestNotaryCertificatesHandlers(t *testing.T) {
@@ -405,525 +391,6 @@ func TestNotaryCertificatesHandlers(t *testing.T) {
 				t.Errorf("expected response did not match.\nExpected vs Received status code: %d vs %d\nExpected vs Received body: \n%s\nvs\n%s\n", tC.status, res.StatusCode, tC.response, string(resBody))
 			}
 		})
-	}
-}
-
-func TestNotaryUsersHandlers(t *testing.T) {
-	testdb, err := db.NewDatabase(":memory:")
-	if err != nil {
-		log.Fatalf("couldn't create test sqlite db: %s", err)
-	}
-	env := &server.Environment{}
-	env.DB = testdb
-	ts := httptest.NewTLSServer(server.NewNotaryRouter(env))
-	defer ts.Close()
-
-	client := ts.Client()
-
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare user accounts and tokens", prepareUserAccounts(ts.URL, client, &adminToken, &nonAdminToken))
-
-	testCases := []struct {
-		desc     string
-		method   string
-		path     string
-		data     string
-		auth     string
-		response string
-		status   int
-	}{
-		{
-			desc:     "Retrieve admin user success",
-			method:   "GET",
-			path:     "/api/v1/accounts/1",
-			data:     "",
-			auth:     adminToken,
-			response: "{\"id\":1,\"username\":\"testadmin\",\"permissions\":1}",
-			status:   http.StatusOK,
-		},
-		{
-			desc:     "Retrieve admin user fail",
-			method:   "GET",
-			path:     "/api/v1/accounts/1",
-			data:     "",
-			auth:     nonAdminToken,
-			response: "error: forbidden",
-			status:   http.StatusForbidden,
-		},
-		{
-			desc:     "Create no password user success",
-			method:   "POST",
-			path:     "/api/v1/accounts",
-			data:     noPasswordUser,
-			auth:     adminToken,
-			response: "{\"id\":3,\"password\":",
-			status:   http.StatusCreated,
-		},
-		{
-			desc:     "Retrieve normal user success",
-			method:   "GET",
-			path:     "/api/v1/accounts/2",
-			data:     "",
-			auth:     adminToken,
-			response: "{\"id\":2,\"username\":\"testuser\",\"permissions\":0}",
-			status:   http.StatusOK,
-		},
-		{
-			desc:     "Retrieve user failure",
-			method:   "GET",
-			path:     "/api/v1/accounts/300",
-			data:     "",
-			auth:     adminToken,
-			response: "error: id not found",
-			status:   http.StatusNotFound,
-		},
-		{
-			desc:     "Create user failure",
-			method:   "POST",
-			path:     "/api/v1/accounts",
-			data:     invalidUser,
-			auth:     adminToken,
-			response: "error: Username is required",
-			status:   http.StatusBadRequest,
-		},
-		{
-			desc:     "Change password success",
-			method:   "POST",
-			path:     "/api/v1/accounts/1/change_password",
-			data:     adminUserNewPassword,
-			auth:     adminToken,
-			response: "1",
-			status:   http.StatusOK,
-		},
-		{
-			desc:     "Change password failure no user",
-			method:   "POST",
-			path:     "/api/v1/accounts/100/change_password",
-			data:     adminUserNewPassword,
-			auth:     adminToken,
-			response: "id not found",
-			status:   http.StatusNotFound,
-		},
-		{
-			desc:     "Change password failure missing password",
-			method:   "POST",
-			path:     "/api/v1/accounts/1/change_password",
-			data:     userMissingPassword,
-			auth:     adminToken,
-			response: "Password is required",
-			status:   http.StatusBadRequest,
-		},
-		{
-			desc:     "Change password failure bad password",
-			method:   "POST",
-			path:     "/api/v1/accounts/1/change_password",
-			data:     userNewInvalidPassword,
-			auth:     adminToken,
-			response: "Password must have 8 or more characters, must include at least one capital letter, one lowercase letter, and either a number or a symbol.",
-			status:   http.StatusBadRequest,
-		},
-		{
-			desc:     "Delete user success",
-			method:   "DELETE",
-			path:     "/api/v1/accounts/2",
-			data:     invalidUser,
-			auth:     adminToken,
-			response: "1",
-			status:   http.StatusAccepted,
-		},
-		{
-			desc:     "Delete user failure",
-			method:   "DELETE",
-			path:     "/api/v1/accounts/2",
-			data:     invalidUser,
-			auth:     adminToken,
-			response: "error: id not found",
-			status:   http.StatusNotFound,
-		},
-	}
-	for _, tC := range testCases {
-		t.Run(tC.desc, func(t *testing.T) {
-			req, err := http.NewRequest(tC.method, ts.URL+tC.path, strings.NewReader(tC.data))
-			req.Header.Add("Authorization", "Bearer "+tC.auth)
-			if err != nil {
-				t.Fatal(err)
-			}
-			res, err := client.Do(req)
-			if err != nil {
-				t.Fatal(err)
-			}
-			resBody, err := io.ReadAll(res.Body)
-			res.Body.Close()
-			if err != nil {
-				t.Fatal(err)
-			}
-			if res.StatusCode != tC.status || !strings.Contains(string(resBody), tC.response) {
-				t.Errorf("expected response did not match.\nExpected vs Received status code: %d vs %d\nExpected vs Received body: \n%s\nvs\n%s\n", tC.status, res.StatusCode, tC.response, string(resBody))
-			}
-			if tC.desc == "Create no password user success" {
-				match, _ := regexp.MatchString(`"password":"[!-~]{16}"`, string(resBody))
-				if !match {
-					t.Errorf("password does not match expected format or length: got %s", string(resBody))
-				}
-			}
-		})
-	}
-}
-
-func TestLogin(t *testing.T) {
-	testdb, err := db.NewDatabase(":memory:")
-	if err != nil {
-		log.Fatalf("couldn't create test sqlite db: %s", err)
-	}
-	env := &server.Environment{}
-	env.DB = testdb
-	env.JWTSecret = []byte("secret")
-	ts := httptest.NewTLSServer(server.NewNotaryRouter(env))
-	defer ts.Close()
-
-	client := ts.Client()
-
-	testCases := []struct {
-		desc     string
-		method   string
-		path     string
-		data     string
-		response string
-		status   int
-	}{
-		{
-			desc:     "Create admin user",
-			method:   "POST",
-			path:     "/api/v1/accounts",
-			data:     adminUser,
-			response: "{\"id\":1}",
-			status:   http.StatusCreated,
-		},
-		{
-			desc:     "Login success",
-			method:   "POST",
-			path:     "/login",
-			data:     adminUser,
-			response: "",
-			status:   http.StatusOK,
-		},
-		{
-			desc:     "Login failure missing username",
-			method:   "POST",
-			path:     "/login",
-			data:     invalidUser,
-			response: "Username is required",
-			status:   http.StatusBadRequest,
-		},
-		{
-			desc:     "Login failure missing password",
-			method:   "POST",
-			path:     "/login",
-			data:     noPasswordUser,
-			response: "Password is required",
-			status:   http.StatusBadRequest,
-		},
-		{
-			desc:     "Login failure invalid password",
-			method:   "POST",
-			path:     "/login",
-			data:     adminUserWrongPass,
-			response: "error: The username or password is incorrect. Try again.",
-			status:   http.StatusUnauthorized,
-		},
-		{
-			desc:     "Login failure invalid username",
-			method:   "POST",
-			path:     "/login",
-			data:     notExistingUser,
-			response: "error: The username or password is incorrect. Try again.",
-			status:   http.StatusUnauthorized,
-		},
-	}
-	for _, tC := range testCases {
-		t.Run(tC.desc, func(t *testing.T) {
-			req, err := http.NewRequest(tC.method, ts.URL+tC.path, strings.NewReader(tC.data))
-			if err != nil {
-				t.Fatal(err)
-			}
-			res, err := client.Do(req)
-			if err != nil {
-				t.Fatal(err)
-			}
-			resBody, err := io.ReadAll(res.Body)
-			res.Body.Close()
-			if err != nil {
-				t.Fatal(err)
-			}
-			if res.StatusCode != tC.status || !strings.Contains(string(resBody), tC.response) {
-				t.Errorf("expected response did not match.\nExpected vs Received status code: %d vs %d\nExpected vs Received body: \n%s\nvs\n%s\n", tC.status, res.StatusCode, tC.response, string(resBody))
-			}
-			if tC.desc == "Login success" && res.StatusCode == http.StatusOK {
-				token, parseErr := jwt.Parse(string(resBody), func(token *jwt.Token) (interface{}, error) {
-					if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-						return nil, fmt.Errorf("Unexpected signing method: %v", token.Header["alg"])
-					}
-					return []byte(env.JWTSecret), nil
-				})
-				if parseErr != nil {
-					t.Errorf("Error parsing JWT: %v", parseErr)
-					return
-				}
-
-				if claims, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
-					if claims["username"] != "testadmin" {
-						t.Errorf("Username found in JWT does not match expected value.")
-					} else if int(claims["permissions"].(float64)) != 1 {
-						t.Errorf("Permissions found in JWT does not match expected value.")
-					}
-				} else {
-					t.Errorf("Invalid JWT token or JWT claims are not readable")
-				}
-			}
-		})
-	}
-}
-
-func TestAuthorization(t *testing.T) {
-	testdb, err := db.NewDatabase(":memory:")
-	if err != nil {
-		log.Fatalf("couldn't create test sqlite db: %s", err)
-	}
-	env := &server.Environment{}
-	env.DB = testdb
-	env.JWTSecret = []byte("secret")
-	ts := httptest.NewTLSServer(server.NewNotaryRouter(env))
-	defer ts.Close()
-
-	client := ts.Client()
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare user accounts and tokens", prepareUserAccounts(ts.URL, client, &adminToken, &nonAdminToken))
-
-	testCases := []struct {
-		desc     string
-		method   string
-		path     string
-		data     string
-		auth     string
-		response string
-		status   int
-	}{
-		{
-			desc:     "metrics reachable without auth",
-			method:   "GET",
-			path:     "/metrics",
-			data:     "",
-			auth:     "",
-			response: "# HELP certificate_requests Total number of certificate requests",
-			status:   http.StatusOK,
-		},
-		{
-			desc:     "status reachable without auth",
-			method:   "GET",
-			path:     "/status",
-			data:     "",
-			auth:     "",
-			response: "",
-			status:   http.StatusOK,
-		},
-		{
-			desc:     "missing endpoints produce 404",
-			method:   "GET",
-			path:     "/this/path/does/not/exist",
-			data:     "",
-			auth:     nonAdminToken,
-			response: "",
-			status:   http.StatusNotFound,
-		},
-		{
-			desc:     "nonadmin can't see accounts",
-			method:   "GET",
-			path:     "/api/v1/accounts",
-			data:     "",
-			auth:     nonAdminToken,
-			response: "",
-			status:   http.StatusForbidden,
-		},
-		{
-			desc:     "admin can see accounts",
-			method:   "GET",
-			path:     "/api/v1/accounts",
-			data:     "",
-			auth:     adminToken,
-			response: `[{"id":1,"username":"testadmin","permissions":1},{"id":2,"username":"testuser","permissions":0}]`,
-			status:   http.StatusOK,
-		},
-		{
-			desc:     "nonadmin can't delete admin account",
-			method:   "DELETE",
-			path:     "/api/v1/accounts/1",
-			data:     "",
-			auth:     nonAdminToken,
-			response: "",
-			status:   http.StatusForbidden,
-		},
-		{
-			desc:     "user can't change admin password",
-			method:   "POST",
-			path:     "/api/v1/accounts/1/change_password",
-			data:     `{"password":"Pwnd123!"}`,
-			auth:     nonAdminToken,
-			response: "",
-			status:   http.StatusForbidden,
-		},
-		{
-			desc:     "user can change self password with /me",
-			method:   "POST",
-			path:     "/api/v1/accounts/me/change_password",
-			data:     `{"password":"BetterPW1!"}`,
-			auth:     nonAdminToken,
-			response: "",
-			status:   http.StatusOK,
-		},
-		{
-			desc:     "user can login with new password",
-			method:   "POST",
-			path:     "/login",
-			data:     `{"username":"testuser","password":"BetterPW1!"}`,
-			auth:     nonAdminToken,
-			response: "",
-			status:   http.StatusOK,
-		},
-		{
-			desc:     "admin can't delete itself",
-			method:   "DELETE",
-			path:     "/api/v1/accounts/1",
-			data:     "",
-			auth:     adminToken,
-			response: "error: deleting an Admin account is not allowed.",
-			status:   http.StatusBadRequest,
-		},
-		{
-			desc:     "admin can delete nonuser",
-			method:   "DELETE",
-			path:     "/api/v1/accounts/2",
-			data:     "",
-			auth:     adminToken,
-			response: "1",
-			status:   http.StatusAccepted,
-		},
-	}
-	for _, tC := range testCases {
-		t.Run(tC.desc, func(t *testing.T) {
-			req, err := http.NewRequest(tC.method, ts.URL+tC.path, strings.NewReader(tC.data))
-			req.Header.Add("Authorization", "Bearer "+tC.auth)
-			if err != nil {
-				t.Fatal(err)
-			}
-			res, err := client.Do(req)
-			if err != nil {
-				t.Fatal(err)
-			}
-			resBody, err := io.ReadAll(res.Body)
-			res.Body.Close()
-			if err != nil {
-				t.Fatal(err)
-			}
-			if res.StatusCode != tC.status || !strings.Contains(string(resBody), tC.response) {
-				t.Errorf("expected response did not match.\nExpected vs Received status code: %d vs %d\nExpected vs Received body: \n%s\nvs\n%s\n", tC.status, res.StatusCode, tC.response, string(resBody))
-			}
-			if tC.desc == "Create no password user success" {
-				match, _ := regexp.MatchString(`"password":"[!-~]{16}"`, string(resBody))
-				if !match {
-					t.Errorf("password does not match expected format or length: got %s", string(resBody))
-				}
-			}
-		})
-	}
-}
-
-func prepareUserAccounts(url string, client *http.Client, adminToken, nonAdminToken *string) func(*testing.T) {
-	return func(t *testing.T) {
-		req, err := http.NewRequest("POST", url+"/api/v1/accounts", strings.NewReader(adminUser))
-		if err != nil {
-			t.Fatal(err)
-		}
-		res, err := client.Do(req)
-		if err != nil {
-			t.Fatal(err)
-		}
-		_, err = io.ReadAll(res.Body)
-		res.Body.Close()
-		if err != nil {
-			t.Fatal(err)
-		}
-		if res.StatusCode != http.StatusCreated {
-			t.Fatalf("creating the first request should succeed when unauthorized. status code received: %d", res.StatusCode)
-		}
-		req, err = http.NewRequest("POST", url+"/api/v1/accounts", strings.NewReader(validUser))
-		if err != nil {
-			t.Fatal(err)
-		}
-		res, err = client.Do(req)
-		if err != nil {
-			t.Fatal(err)
-		}
-		_, err = io.ReadAll(res.Body)
-		res.Body.Close()
-		if err != nil {
-			t.Fatal(err)
-		}
-		if res.StatusCode != http.StatusUnauthorized {
-			t.Fatalf("the second request should have been rejected. status code received: %d", res.StatusCode)
-		}
-		req, err = http.NewRequest("POST", url+"/login", strings.NewReader(adminUser))
-		if err != nil {
-			t.Fatal(err)
-		}
-		res, err = client.Do(req)
-		if err != nil {
-			t.Fatal(err)
-		}
-		resBody, err := io.ReadAll(res.Body)
-		res.Body.Close()
-		if err != nil {
-			t.Fatal(err)
-		}
-		if res.StatusCode != http.StatusOK {
-			t.Fatalf("the admin login request should have succeeded. status code received: %d", res.StatusCode)
-		}
-		*adminToken = string(resBody)
-		req, err = http.NewRequest("POST", url+"/api/v1/accounts", strings.NewReader(validUser))
-		req.Header.Set("Authorization", "Bearer "+*adminToken)
-		if err != nil {
-			t.Fatal(err)
-		}
-		res, err = client.Do(req)
-		if err != nil {
-			t.Fatal(err)
-		}
-		_, err = io.ReadAll(res.Body)
-		res.Body.Close()
-		if err != nil {
-			t.Fatal(err)
-		}
-		if res.StatusCode != http.StatusCreated {
-			t.Fatalf("creating the second request should have succeeded when given the admin auth header. status code received: %d", res.StatusCode)
-		}
-		req, err = http.NewRequest("POST", url+"/login", strings.NewReader(validUser))
-		if err != nil {
-			t.Fatal(err)
-		}
-		res, err = client.Do(req)
-		if err != nil {
-			t.Fatal(err)
-		}
-		resBody, err = io.ReadAll(res.Body)
-		res.Body.Close()
-		if err != nil {
-			t.Fatal(err)
-		}
-		if res.StatusCode != http.StatusOK {
-			t.Errorf("the admin login request should have succeeded. status code received: %d", res.StatusCode)
-		}
-		*nonAdminToken = string(resBody)
 	}
 }
